refactor(utils): parseInput, validateData, printData 파일 위치 변경
